[ONL-8214] chore: Re-exported types directly.
