+++ global.d.ts
@@ -5,12 +5,3 @@ declare module 'vue' {
 }
 
 export type Maybe<T> = T | null;
-
-export enum ZIndexLevel {
-  NOTIFICATION = 'notification',
-  MODAL = 'modal',
-  TOOLTIP = 'tooltip',
-  LEVEL1 = 'level-1',
-  LEVEL2 = 'level-2',
-  LEVEL3 = 'level-3',
-}
