+++ src/components/ec-alert/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-alert.vue';
-export * as types from './types';
