+++ src/components/ec-btn/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-btn.vue';
-export * as types from './types';
