+++ src/components/ec-button-group/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-button-group.vue';
-export * as types from './types';
