+++ src/components/ec-checkbox/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-checkbox.vue';
-export * as types from './types';
