+++ src/components/ec-icon/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-icon.vue';
-export * as types from './types';
