+++ src/components/ec-input-field/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-input-field.vue';
-export * as types from './types';
