+++ src/components/ec-loading/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-loading.vue';
-export * as types from './types';
