+++ src/components/ec-main-container/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-main-container.vue';
-export * as types from './types';
