+++ src/components/ec-metroline/components/ec-metroline-item/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-metroline-item.vue';
-export * as types from './types';
