+++ src/components/ec-metroline/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-metroline.vue';
-export * as types from './types';
