+++ src/components/ec-metroline/types.ts
@@ -1,4 +1,4 @@
-export * from './provide';
+import type { Maybe } from '../../../global';
 
 export enum MetrolineEvent {
   CHANGE = 'change',
@@ -8,4 +8,15 @@ export enum MetrolineEvent {
 export interface MetrolineEvents {
   [MetrolineEvent.CHANGE]: number
   [MetrolineEvent.COMPLETE]: undefined
+}
+
+export interface MetrolineProviderContext {
+  activeItemId: Maybe<number>,
+  lastItemId: Maybe<number>,
+  isCompleted: boolean,
+  register: (id: number) => void,
+  unregister: (id: number) => void,
+  goToNext: (id: number) => void,
+  goTo: (id: number) => void,
+  complete: () => void,
 }
