+++ src/components/ec-navigation-arrows/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-navigation-arrows.vue';
-export * as types from './types';
