+++ src/components/ec-option-card/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-option-card.vue';
-export * as types from './types';
