+++ src/components/ec-phone-number-input/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-phone-number-input.vue';
-export * as types from './types';
