+++ src/components/ec-radio-btn-group/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-radio-btn-group.vue';
-export * as types from './types';
