+++ src/components/ec-radio-btn/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-radio-btn.vue';
-export * as types from './types';
