+++ src/components/ec-table/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-table.vue';
-export * as types from './types';
