+++ src/components/ec-tag/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-tag.vue';
-export * as types from './types';
