+++ src/components/ec-text-filter/index.ts
@@ -1,3 +1,2 @@
 export { default } from './ec-text-filter.vue';
-export * as types from './types';
