+++ src/composables/use-ec-sorting/index.ts
@@ -1,2 +1,1 @@
-export * as types from './types';
 export { default } from './use-ec-sorting';
