+++ src/directives/ec-amount/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-amount';
-export * as types from './types';
