+++ src/directives/ec-tooltip/index.ts
@@ -1,2 +1,1 @@
 export { default } from './ec-tooltip';
-export * as types from './types';
