+++ src/enums/index.ts
@@ -13,3 +13,12 @@ export enum KeyCode {
   ARROW_UP = 38,
   ARROW_DOWN = 40,
 }
+
+export enum ZIndexLevel {
+  NOTIFICATION = 'notification',
+  MODAL = 'modal',
+  TOOLTIP = 'tooltip',
+  LEVEL1 = 'level-1',
+  LEVEL2 = 'level-2',
+  LEVEL3 = 'level-3',
+}
