+++ src/icons/currency-flags/index.ts
@@ -1,2 +1,1 @@
 export { availableCurrencyFlagsSet, flags } from './currency-flags';
-export * as types from './types';
