+++ src/main.ts
@@ -1,11 +1,16 @@
-export { default as EcAlert, types as EcAlertTypes } from './components/ec-alert';
+export type * from '../global';
+export { default as EcAlert } from './components/ec-alert';
+export * from './components/ec-alert/types';
 export { default as EcAmountFilterInput } from './components/ec-amount-filter-input';
 export { default as EcAmountInput } from './components/ec-amount-input';
 export { default as EcBadge } from './components/ec-badge';
-export { default as EcBtn, types as EcBtnTypes } from './components/ec-btn';
+export { default as EcBtn } from './components/ec-btn';
+export * from './components/ec-btn/types';
 export { default as EcBtnDropdown } from './components/ec-btn-dropdown';
-export { default as EcButtonGroup, types as EcButtonGroupTypes } from './components/ec-button-group';
+export { default as EcButtonGroup } from './components/ec-button-group';
+export * from './components/ec-button-group/types';
 export { default as EcCheckbox } from './components/ec-checkbox';
+export * from './components/ec-checkbox/types';
 export { default as EcContainer } from './components/ec-container';
 export { default as EcCurrencyFilter } from './components/ec-currency-filter';
 export { default as EcCurrencyInput } from './components/ec-currency-input';
@@ -19,38 +24,54 @@ export { default as EcFileList } from './components/ec-file-list';
 export { default as EcFileUpload } from './components/ec-file-upload';
 export { default as EcFilterPopover } from './components/ec-filter-popover';
 export { default as EcFullScreenOverlay } from './components/ec-full-screen-overlay';
-export { default as EcIcon, types as EcIconTypes } from './components/ec-icon';
+export { default as EcIcon } from './components/ec-icon';
+export * from './components/ec-icon/types';
 export { default as EcInlineActions } from './components/ec-inline-actions';
 export { default as EcInlineInputField } from './components/ec-inline-input-field';
-export { default as EcInputField, types as EcInputFieldTypes } from './components/ec-input-field';
-export { default as EcLoading, types as EcLoadingTypes } from './components/ec-loading';
+export * from './components/ec-inline-input-field/types';
+export { default as EcInputField } from './components/ec-input-field';
+export * from './components/ec-input-field/types';
+export { default as EcLoading } from './components/ec-loading';
+export * from './components/ec-loading/types';
 export { default as EcLoadingIcon } from './components/ec-loading-icon';
-export { default as EcMainContainer, types as EcMainContainerTypes } from './components/ec-main-container';
+export { default as EcMainContainer } from './components/ec-main-container';
+export * from './components/ec-main-container/types';
 export { default as EcMenu } from './components/ec-menu';
-export { default as EcMetroline, types as EcMetrolineTypes } from './components/ec-metroline';
-export { default as EcMetrolineItem, types as EcMetrolineItemTypes } from './components/ec-metroline/components/ec-metroline-item';
+export { default as EcMetroline } from './components/ec-metroline';
+export { default as EcMetrolineItem } from './components/ec-metroline/components/ec-metroline-item';
+export * from './components/ec-metroline/components/ec-metroline-item/types';
 export { METROLINE_PROVIDE_KEY } from './components/ec-metroline/provide';
+export * from './components/ec-metroline/types';
 export { default as EcModal } from './components/ec-modal';
 export { default as EcMultipleValuesSelection } from './components/ec-multiple-values-selection';
 export { default as EcNavigation } from './components/ec-navigation';
-export { default as EcNavigationArrows, types as EcNavigationArrowsTypes } from './components/ec-navigation-arrows';
+export { default as EcNavigationArrows } from './components/ec-navigation-arrows';
+export * from './components/ec-navigation-arrows/types';
 export { default as EcNavigationLink } from './components/ec-navigation-link';
-export { default as EcOptionCard, types as EcOptionCardTypes } from './components/ec-option-card';
+export { default as EcOptionCard } from './components/ec-option-card';
+export * from './components/ec-option-card/types';
 export { default as EcPanel } from './components/ec-panel';
-export { default as EcPhoneNumberInput, types as EcPhoneNumberInputTypes } from './components/ec-phone-number-input';
+export { default as EcPhoneNumberInput } from './components/ec-phone-number-input';
+export * from './components/ec-phone-number-input/types';
 export { default as EcPopover } from './components/ec-popover';
 export { POPOVER_CONTAINER_KEY } from './components/ec-popover/ec-popover-provide';
 export { default as EcPrivacyPolicy } from './components/ec-privacy-policy';
-export { default as EcRadioBtn, types as EcRadioBtnTypes } from './components/ec-radio-btn';
-export { default as EcRadioBtnGroup, types as EcRadioBtnGroupTypes } from './components/ec-radio-btn-group';
+export { default as EcRadioBtn } from './components/ec-radio-btn';
+export * from './components/ec-radio-btn/types';
+export { default as EcRadioBtnGroup } from './components/ec-radio-btn-group';
+export * from './components/ec-radio-btn-group/types';
 export { default as EcSmartTable } from './components/ec-smart-table';
 export { default as EcSubmenu } from './components/ec-submenu';
 export { default as EcSummaryInfo } from './components/ec-summary-info';
+export * from './components/ec-summary-info/types';
 export { default as EcSyncMultipleValuesFilter } from './components/ec-sync-multiple-values-filter';
-export { default as EcTable, types as EcTableTypes } from './components/ec-table';
+export { default as EcTable } from './components/ec-table';
+export * from './components/ec-table/types';
 export { default as EcTableFilter } from './components/ec-table-filter';
-export { default as EcTag, types as EcTagTypes } from './components/ec-tag';
-export { default as EcTextFilter, types as EcTextFilterTypes } from './components/ec-text-filter';
+export { default as EcTag } from './components/ec-tag';
+export * from './components/ec-tag/types';
+export { default as EcTextFilter } from './components/ec-text-filter';
+export * from './components/ec-text-filter/types';
 export { default as EcTextarea } from './components/ec-textarea';
 export { default as EcTimer } from './components/ec-timer';
 export { default as EcToaster } from './components/ec-toaster';
@@ -62,15 +83,16 @@ export { default as useEcSorting } from './composables/use-ec-sorting';
 export * from './composables/use-ec-sorting/types';
 export type { EcConfig } from './config';
 export { default as config } from './config';
-export { default as EcAmount, default as VEcAmount } from './directives/ec-amount';
-export { default as EcClosePopover, default as VEcClosePopover } from './directives/ec-close-popover';
-export {
-  default as EcTooltip, default as VEcTooltip, default as vEcTooltip, types as VEcTooltipTypes,
-} from './directives/ec-tooltip';
+export { default as EcAmount, default as VEcAmount, default as vEcAmount } from './directives/ec-amount';
+export * from './directives/ec-amount/types';
+export { default as EcClosePopover, default as VEcClosePopover, default as vEcClosePopover } from './directives/ec-close-popover';
+export { default as EcTooltip, default as VEcTooltip, default as vEcTooltip } from './directives/ec-tooltip';
+export * from './directives/ec-tooltip/types';
 export * from './enums';
 export { default as ecWithAbortableFetch } from './hocs/ec-with-abortable-fetch';
 export { createHOC, createHOCc } from './hocs/hoc';
 export { inlineSvgSprites } from './icons/browser';
-export { availableCurrencyFlagsSet, types as AvailableCurrencyFlagsTypes, flags } from './icons/currency-flags';
+export { availableCurrencyFlagsSet, flags } from './icons/currency-flags';
+export * from './icons/currency-flags/types';
 export * from './icons/types';
 export { default as flatpickr } from 'flatpickr';
